format

+++ src/main/java/dev/xkmc/cuisine/init/data/WoodType.java
@@ -271,7 +271,7 @@ class CustomBoatItem extends Item {
 				Vec3 vec31 = p_40623_.getEyePosition();
 
 				for (Entity entity : list) {
-					AABB aabb = entity.getBoundingBox().inflate((double) entity.getPickRadius());
+					AABB aabb = entity.getBoundingBox().inflate(entity.getPickRadius());
 					if (aabb.contains(vec31)) {
 						return InteractionResultHolder.pass(itemstack);
 					}
